Pass text style generation to a mixin
